feat: Add ids for in-page linking
To allow linking to specific areas of page content. e.g. #where on the
classes page to link to the section detailing the location of the club.
In order to allow this to link to the correct area I have placed the
links onto separate divs positioned to take into account the height of
the fixed header. However they are actually sitting at the bottom, in
the content areas above the ones they relate to rather than their actual
content areas. I have read that this is not the best semantically, so
it might be better to find a solution that allows these divs to be
inside the correct place.